+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased] 1.0.0-alpha.5
 ### Fixed
 - Passing start date when checking availability no longer makes that date the primary rules start date.
+- Bookings are unique per-slot
+
+### Improved
+- Querying availability in range now much faster
 
 ## Added
 - Added `Event::getNextAvailableSlot()`